removed all files in /dist and renamed main.js to index.js
