+++ package.json.old
@@ -1,0 +1,38 @@
+{
+  "name": "rpg-game",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "start": "parcel src/index.html -p 5555",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/maosan132/rpg-game.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/maosan132/rpg-game/issues"
+  },
+  "homepage": "https://github.com/maosan132/rpg-game#readme",
+  "dependencies": {
+    "assets": "^3.0.1",
+    "parcel-bundler": "^1.12.5",
+    "phaser": "^3.54.0"
+  },
+  "devDependencies": {
+    "babel-eslint": "10.1.x",
+    "eslint": "6.8.x",
+    "eslint-config-airbnb-base": "14.1.x",
+    "eslint-plugin-import": "2.20.x",
+    "stylelint": "13.3.x",
+    "stylelint-config-standard": "20.0.x",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "3.17.x",
+    "webpack": "^5.45.1",
+    "webpack-cli": "^4.7.2"
+  }
+}
